requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
python-dotenv==1.0.0
tiktoken>=0.5.2,<0.6.0
fastapi==0.110.2
uvicorn==0.29.0 